example_app/backend/requirements.txt
openai==0.27.4
python-dotenv==1.0.0
websockets==11.0.3
-e ../../llmtaskgraph